requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
future==0.18.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydub==0.25.1
pyee==11.0.0
python-dateutil==2.8.2
python-ffmpeg==2.0.4
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
termcolor==2.3.0
typing-extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
